// apps/api/src/modules/auth/auth.validators.ts
import { z } from 'zod';

export const registerSchema = z.object({
    body: z.object({
        username: z.string().min(3).max(30),
        email: z.string().email(),
        password: z.string().min(8),
    }),
});

export const loginSchema = z.object({
    body: z.object({
        email: z.string().email(),
        password: z.string().min(1),
    }),
});